Add unix authentication
Add the nfs.AuthenticationHandler to be configured in the
BackendSession. If the client authentication is accepted, the
credentials are passed to the fs.FS via the new SetCreds function. If
the client authentication is denied, a proper rpc answer is sent to the
client.

Since the SetCreds and the other FS functions should not interleave for
different sessions, we now create a new FS for each connection in the
provided backend. This could be avoided if instead of a SetCreds
function, credentials are passed as argument to all FS' functions.

Signed-off-by: Peter Verraedt <peter@verraedt.be>